- 提交，回家继续搞

+++ addons/LdtkImporter/LdtkJson.cs
@@ -1158,7 +1158,7 @@ namespace Ldtk
         public long? TagsSourceEnumUid { get; set; }
 
         [JsonPropertyName("tileGridSize")]
-        public long TileGridSize { get; set; }
+        public int TileGridSize { get; set; }
 
         /// <summary>
         /// Unique Intidentifier
